server/requirements.txt
pysqlite3  # This library takes the SQLite module from Python 3 and packages it as a separately-installable module
python-multipart  # Not sure what this is for
mysql-connector-python  # MySQL driver written in Python which does not depend on MySQL C client libraries and implements the DB API v2.0 specification (PEP-249).
fastapi==0.85.1
SQLAlchemy~=2.0.23
python-dotenv
pyyaml